feat: complete final tast module3 v.1

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        double temperatureFahrenheit = (temperatureCelsius*9/5) + 32;
+        double temperatureFahrenheit = (double) temperatureCelsius*9/5 + 32;
         System.out.println(temperatureFahrenheit);
     }
 }
